+++ HISTORY.rst
@@ -13,8 +13,46 @@ Releases
 ********
 
 
+3.8.0 (2021-03-30)
+__________________
+
+New Features
+------------
+* Watershed partitioning now supports dask (`PR27 <https://github.com/wavespectra/wavespectra/pull/27>`_).
+* Spectral splitting now supports dask.
+* The following spectral parameters now support dask (`PR11 <https://github.com/wavespectra/wavespectra/pull/11>`_):
+
+  * tp
+  * dp
+  * dpm
+  * dspr
+* Wavespectra conda recipe by `RubendeBruin`_.
+
+Internal Changes
+----------------
+* Core watershed partitioning code organised into watershed module.
+* `max_swells` replaced by `swells` in watershed partition to return fixed number of swells.
+* Renamed module `wavespectra.core.misc` by `wavespectra.core.utils`.
+* Removed deprecated method `_same_dims`, `_inflection` and `_product` from `SpecArray`.
+* Get rid of simpy dependency.
+* New daskable stats defined as ufuncs using numba.
+* SpecArray attributes redefined as property methods.
+
+Bug Fixes
+---------
+
+deprecation
+-----------
+* Drop support for python < 3.7
+* Dropped args `hs_min` and `nearest` in `SpecArray.partition`.
+
+
+.. _`RubendeBruin`: https://github.com/RubendeBruin
+
+
 3.7.2 (2021-01-12)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 New Features
 ------------
@@ -37,7 +75,8 @@ Bug Fixes
 
 
 3.7.1 (2020-08-26)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 Internal Changes
 ----------------
@@ -50,7 +89,8 @@ Internal Changes
 
 
 3.7.0 (2020-07-16)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 New Features
 ------------
@@ -62,7 +102,8 @@ Internal Changes
 
 
 3.6.5 (2020-07-10)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 Bug Fixes
 ---------
@@ -70,7 +111,8 @@ Bug Fixes
 
 
 3.6.4 (2020-06-29)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 Bug Fixes
 ---------
@@ -78,7 +120,8 @@ Bug Fixes
 
 
 3.6.3 (2020-06-28)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 Internal Changes
 ----------------
@@ -86,7 +129,8 @@ Internal Changes
 
 
 3.6.2 (2020-06-28)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 Internal Changes
 ----------------
@@ -94,7 +138,8 @@ Internal Changes
 
 
 3.6.1 (2020-06-28)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 Internal Changes
 ----------------
@@ -102,7 +147,8 @@ Internal Changes
 
 
 3.6.0 (2020-06-27)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 New Features
 ------------
@@ -134,7 +180,8 @@ Internal Changes
 
 
 3.5.3 (2020-04-14)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 Fix xarray version until breaking changes with 0.15.1 are taken care of.
 
 Bug Fixes
@@ -147,7 +194,8 @@ Internal Changes
 
 
 3.5.2 (2020-03-09)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 New Features
 ------------
@@ -160,7 +208,8 @@ New Features
 
 
 3.5.1 (2019-12-12)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 Bug Fixes
 ---------
@@ -173,7 +222,8 @@ Internal Changes
 
 
 3.5.0 (2019-12-09)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 **The first PyPI release from new** `wavespectra`_ **github organisation.**
 
 Breaking Changes
@@ -214,7 +264,8 @@ Internal Changes
 
 
 3.4.0 (2019-03-28)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 **The last PyPI release from old metocean github organisation.**
 
 New Features
@@ -223,7 +274,8 @@ New Features
 
 
 3.3.1 (2019-03-19)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 New Features
 ------------
@@ -232,7 +284,8 @@ New Features
 
 
 3.3.0 (2019-02-21)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 New Features
 ------------
@@ -248,7 +301,8 @@ Internal Changes
 
 
 3.2.5 (2019-01-25)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 Bug Fixes
 ---------
@@ -256,7 +310,8 @@ Bug Fixes
 
 
 3.2.4 (2019-01-23)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 Bug Fixes
 ---------
@@ -264,7 +319,8 @@ Bug Fixes
 
 
 3.2.3 (2019-01-08)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 New Features
 ------------
@@ -276,7 +332,8 @@ Bug Fixes
 
 
 3.2.2 (2018-12-04)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 Bug Fixes
 ---------
@@ -284,7 +341,8 @@ Bug Fixes
 
 
 3.2.1 (2018-11-14)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 New Features
 ------------
@@ -296,7 +354,8 @@ Bug Fixes
 
 
 3.2.0 (2018-11-04)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 New Features
 ------------
@@ -312,7 +371,8 @@ Internal changes
 
 
 3.1.4 (2018-08-29)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 Bug Fixes
 ---------
@@ -320,7 +380,8 @@ Bug Fixes
 
 
 3.1.3 (2018-07-27)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 Changes
 -------
@@ -328,7 +389,8 @@ Changes
 
 
 3.1.2 (2018-07-05)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 Changes
 -------
@@ -340,7 +402,8 @@ Bug Fixes
 
 
 3.1.1 (2018-05-17)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 Bug Fixes
 ---------
@@ -351,7 +414,8 @@ New Features
 
 
 3.1.0 (2018-05-09)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 New Features
 ------------
@@ -363,7 +427,8 @@ Bug Fixes
 
 
 3.0.2 (2018-05-03)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 Bug Fixes
 ---------
@@ -371,7 +436,8 @@ Bug Fixes
 
 
 3.0.1 (2018-04-28)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 
 New Features
 ------------
@@ -395,7 +461,8 @@ Internal Changes
 
 
 3.0 (2018-03-05)
-~~~~~~~~~~~~~~~~~~
+__________________
+
 **This major release marks the migration from the predecessor** `pyspectra` **library,
 as well as the open-sourcing of wavespectra and first PyPI release.**